<fix>[plugin]: config url of zbs
Resolves: ZSTAC-71766
